nvidia-settings: Remove $(STAMP_C)
This file is generated at build time and includes the version, builder, and date
of the build. For example,

 $ strings /usr/bin/nvidia-settings | grep 'nvidia id'
 nvidia id: nvidia-settings:  version 396.14  (buildmeister@swio-display-x86-rhel47-09)  Mon Mar 19 20:18:47 PDT 2018

This information isn't hugely useful, and makes it difficult for some
distributions to achieve their goal of reproducible builds. Rather than adding a
build-time flag to hard-code this information, just remove it.

See https://github.com/NVIDIA/nvidia-settings/pull/12