nvidia-settings: Remove __DATE__ from the man page
The man page is supposed to have a date in it that describes when it was
updated. It's hard to remember to change it so we have the makefile insert the
build date instead. However, doing that makes it difficult for distributions to
achieve reproducible builds since it changes the resulting package when builds
happen on different days.

The date in this file doesn't really matter, so just hard-code today's date and
remove the code to generate it.

Closes https://github.com/NVIDIA/nvidia-settings/pull/12